Moved integration tests out of unit tests.
